Add VSCode Icons and custom css
