+++ package.json
@@ -25,6 +25,7 @@
     "redhat.vscode-yaml"
   ],
   "dependencies": {
+    "@vscode/codicons": "^0.0.33",
     "@vscode/webview-ui-toolkit": "^1.2.2",
     "sf_toolkit": "github:silverfin/sf-toolkit",
     "yaml": "^2.2.2"
