+++ src/lib/firmHandler.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import * as utils from "./utils";
+import * as utils from "../utilities/utils";
 const sfToolkit = require("sf_toolkit");
 const { config } = require("sf_toolkit/api/auth");
