+++ src/lib/media/style.css
@@ -1,0 +1,18 @@
+html {
+  box-sizing: border-box;
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+
+ol,
+ul {
+  list-style: none;
+}
+
+body {
+  padding: 5px 10px;
+}
